+++ pkg/conf/store.go
@@ -97,7 +97,7 @@ func (s *Store) MaybeUpdate(rawConfig string) (UpdateResult, error) {
 
 	s.raw = rawConfig
 
-	newConfig, err := parse.ParseConfigEnvironment(rawConfig)
+	newConfig, err := parse.ParseConfigData(rawConfig)
 	if err != nil {
 		return result, errors.Wrap(err, "when parsing rawConfig during update")
 	}